Move an IDW helper into a new module, as it's not specifically YAML-related
